snps/requirements.txt
defusedxml==0.4.1
Django==1.10.5
django-allauth==0.30.0
django-chunked-upload==1.1.1
django-threadlocals==0.8
django-widget-tweaks==1.4.1
numpy==1.12.0
oauthlib==2.0.1
pandas==0.19.2
pymongo==3.4.0
python-dateutil==2.6.0
python3-openid==3.0.10
pytz==2016.10
requests==2.13.0
requests-oauthlib==0.7.0
six==1.10.0
xlrd==1.0.0
beautifulsoup4==4.5.1
bs4==0.0.1
py==1.4.32
pytest==3.0.5
urllib5
wheel==0.24.0
wikitools==1.3
